// src/lib/api/generated/models/eventSinkKind.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.1
 */

export type EventSinkKind = typeof EventSinkKind[keyof typeof EventSinkKind];


export const EventSinkKind = {
  webhook: 'webhook',
  amqp: 'amqp',
  valkey_stream: 'valkey_stream',
  email: 'email',
} as const;